refactor(api): consolidate response functions, remove dead code
- Move createApiErrorResponse, createApiSuccessResponse, logAndSanitizeError to response.ts
- Remove sanitizeErrorForClient and withApiErrorHandling (dead code)
- Keep createResponseHeaders, logApiRequest, logApiResponse (actively used in contact route)
- Update projects/route.ts imports to use response.ts
